+++ changelog.md
@@ -2,3 +2,6 @@
 Initial Release
 0.1.1
 Fixed infinite recursion bug. Now playable.
+0.2.1
+Added support for story point improvement as well as two colony items:
+Biofactory embryos and pristine nanoforge.